requirements.txt
ccxt
asyncio
websockets
psycopg2-binary
pandas
pandas_ta
fastapi
uvicorn
python-dotenv
openai
